requirements.txt
asgiref==3.2.5
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
django-core==1.4.1
djangorestframework==3.11.1
gunicorn==20.0.4
h5py==2.10.0
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
joblib==0.16.0
numpy==1.19.2
pandas==1.1.2
Pillow==7.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.2
#scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.0
threadpoolctl==2.1.0
urllib3==1.25.10
websocket-client==0.48.0
whitenoise==5.2.0
wincertstore==0.2
